Update electron-builder artifactName to {name}
Our product name contains a space. Github releases doesn't allow
files with spaces and sanitizes the name replacing the spaces with ".".

Electron-builder sanitizes names using a "-" to replace spaces. This
creates a discrepancy between which files electron-updater (the auto update
part of builder) expects to be available for download from the latest
github release.

This change uses "name" which does not have spaces.