Add bag file icon
macOS Big Sur can automatically generate icons using the app icon and overlay text. This requires changes in the Info.plist CFBundleDocumentTypes, but since electron-builder controls this key by default, we have to unset fileAssociations and provide the Info.plist values completely manually for Mac.

The fallback icon should still be used on macOS pre-11, but for some reason the legacy icon was not used when I disabled CFBundleTypeIconSystemGenerated on my machine. Testing on an older OS is probably warranted.