Add player "presence" enum replacing showSpinner & showInitializing
Add a new PlayerPresence enum to player state, and remove showSpinner and showInitializing.

Change player to "maybePlayer" in the MessagePipelineProvider API, allowing the PlayerManager to inform it about player construction status. The MessagePipelineProvider is responsible for merging this information with the player's own presence info (such as initializing/connected/reconnecting) to produce the final `playerState.presence`.

Add a spinner / error indicator into the top bar to show player connection status. Error messages are currently not displayed, but we should do this ASAP.

Also extract MockMessagePipelineProvider into its own file.