Fix webpack build failure detection for integration tests
Integration tests would stall if the webpack build failed with errors
because the _err_ for compiler.run is insufficient to detect build errors.
We also need to check result.hasErrors() to see if the build error'd.

Now the integration test fails when the build fails rather than trying
to launch the app and timing out.