Load the Roboto Mono font on chart.js worker thread explicitly
A large portion of Windows users are affected by a rendering thread crash in [skia](https://skia.org/) that seems to happen when chart.js renders a custom font to a <canvas> element when the system display scaling is set >100%.

This PR directly fixes a related issue, the Roboto Mono font is specified for chart axis labels but is not actually available in the worker thread where `ctx.fillText()` is called. By waiting for the custom font to be loaded before rendering the chart, it also appears to fix the DirectWrite font crash in my testing so far.

Fixes https://github.com/foxglove/studio/issues/495