Restrict `allowDangerousHtml` to specific parts of the UI
allowDangerousHtml was set to true recently to allow HTML formatting in Markdown help pages for panels. This PR rolls the setting back from being universally enabled to conditionally enabling in a few parts of the app, to prevent a future problem if we ever start rendering Markdown from user input (such as embedded in a bag).