Rework Plot downsampling and state management
The downsampling and state management was causing too many renders and re-renders
on the plot when scale bounds changed or when new data was introduced to the dataset.

This change fixes this by using a callback for the scale update from the ChartComponent
and dispatching the scales to globalScales or downsampling accordingly.