Fix handling of inverted min/max ranges from user input.
If a user entered a max value which was smaller than any datapoint value,
chart.js would crash when computing linear scales. We fix this by
normalizing min/max within TimeBasedChart so min < max.

This change also makes it so that when the user removes their manual
override of the y-axis bounds, the chart returns back to dataset bounds.
Unfortunately we need to set the y-axis bounds manually in that case
possibly due to chartjs not re-calculating the bounds when we set min/max
back to undefined.

Fixes #875