Upgrade electron to v13 beta27
electron-chromedriver (used by spectron) downloads a version of chromedriver from electron builds based on the version of the electron-chromedriver package. So if electron-chromedriver is v12 it will download chromedriver for electron v12. Spectron was using electron-chromedriver v12 which didn't support the newer chrome in the recent v13 betas. We override electron-chromedriver for spectron so the correct chromedriver version is downloaded.