Add RemoteLayoutStorage and LayoutStorage interfaces
The RemoteLayoutStorage interface demonstrates what operations a layout storage server with sharing features will be expected to implement. This may change a little in the near future, but has largely solidified after discussions and prototyping work on the server side.

The existing LayoutStorage is renamed to LocalLayoutStorage. It's still the only storage interface being used in the app right now, but will be combined in a future PR with remote storage to form a local cache inplementation.