Fix: Clear panel element before calling initPanel again
When a player changes, it triggers the effect responsible for initPanel
to run. This effect was not properly clearing any elements the previous
initPanel set on the panel element which resulted in orphaned dom elements
persisting in the dom.

This change clears the panel element as part of the prior effect cleanup.