Add share button; remove updateSharedLayout API
Previously we had an updateSharedLayout that required a sourceID. This makes uploading local edits to shared layouts challenging because it requires 2 steps: upload the edits (under a new id?) and then call updateSharedLayout to replace the shared layout with the new changes. Keeping it as one update API that always accepts new data (instead of a source layout id) should be simpler.