Fix: RawMessage panel Copy msg to handle lazy messages, bigints, and typed arrays
Previously, the Copy msg feature of RawMessage would attempt to deep clone
the data before invoking JSON.stringify. This doesn't work for lazy messages as
they are not trivially deep clonable.

This change passes the data directory to JSON.stringify which understands `.toJSON`
functions on objects and uses that if available. Additionally, a replacer is provided
which handles bigint and typed arrays.

Fixes #1432