Feature: allow bags with empty message definitions
An empty message definition is a valid message definition. This
change allows bags that have topics with empty message definitions
to load and play. Since the definition is empty, such topics are
not desirialized.

Alters the fix performed in webviz in response to https://github.com/cruise-automation/webviz/issues/373 which would previously crash the
app when there was an empty message definition.

This change removes a flaky test for the notification when loading such
a bag since it is valid to load such a bag.