Add a disabledReason field to PlayerSourceDefinition
The disabledReason field allows components (like the ConnectionList)
to show an error if the particular player is not available on that
platform.

On web, this allows us to direct the user to download the desktop app
which does have the desired connection.

Change useConfirm to match usePrompt api interfaces and move to hooks folder.