Fix: disable animations on chart updates
Since we replace all data points in the chart when updating (due to
how we send data over the worker), chart animations cause lag for updates
as the chart tries to re-compute animations for all the points.

For our uses we don't need animations so we disable them for the update call
and in the options.