Fix: limit the plot history size in follow mode
When streaming data, the plot would keep an infinite history of
messages even when a fixed follow window of some seconds was set. Keeping
the app conntected would result in the memory usage continuing to grow
and more dataset processing to occur as the messages array grew.

With this change, if there is a follow window, messages older than the
window are discarded.