Fix: plots properly resize when their container size changes
https://github.com/foxglove/studio/pull/1705 introduced a regression
where plots would not resize when their container size changed.

The problem occured because the getNewUpdateMessage logic only includes
changed fields in the update rpc message but the chart worker resize handling
assumed that both width and height would be present for a resize.

This change fixes that assumption and allows the chart to resize when either
width or height change.

Fixes #1714