Remove sendNotification from SceneBuilder
SceneBuilder would send a notification when a topic had a bad frame id. This notification
would popup for the user and persist even tho scene builder might receive more topics
and frames and internally clear the error.

This change removes the sendNotification and uses the already present per-topic
notification mechanism local to the 3d panel to inform the user if there are errors on
their topic. This mechanism properly clears itself when there are no errors on the topic.