Console player: fix preload range calculation
The fullyLoadedExtent function was not well suited for preload calculation because it was unclear what it should return if the target time was the end of a loaded range or equal to the maxTime. This led to issues where an incorrect preload request would be made resulting in attempting to add overlapping data to the cache. Replace it with a nextRangeToLoad function which directly calculates the range we need to load.