Disable the rename data provider
When users load multiple bag files the more common expectation is that these files would load as one giant bag. The existing behavior - inherited from webviz - would try to load the second bag under a prefix with the idea being a user could compare the bags. For now we comment out this feature to see what the impact is for users. We'll aim to remove this feature and re-introduce it when we re-work data source selection to allow for loading multiple bags as one bag or with a prefix.