Use plot defaultConfig when destructuring config values
The default config is only applied when a panel is added to a layout. When we add new config options to the default config, these are not applied to existing panels. Any default options with "truthy" values are thus not actually set to those values for the panel.

This change moves the default value assignment into the panel logic so it is the responsibility of the panel to provide defaults for its options.