Fix message path input valid path detection for time and duration
time and duration are considered primative types in ROS but are actually objects with two fields (sec, nsec). When specifying the sec or nsec field within a time field (i.e. header.stamp.sec) in the message path input, it would be marked as red indicating the path is invalid. This change fixes this behavior so paths accessing .sec or .nsec of time or duration fields is shown as valid.

The change produces a structure for time and duration with the message structure type and contains sec and nsec primitive fields. Previously the structure had time and duration primative fields with no knowledge of sec or nsec below.

Fixes: #1958