Fix CSV plot download for (u)int64 types
When downloading data from the plot, the datum y field was used. This field is the Number conversion from the original y value. The unconverted y value is in the _value_ field. The _value_ field maintains the original bigint type. Use the _value_ field so the original data is present in the csv output.

Fixes: #2183