Toggle fullscreen state for parent panels
When a panel within a tab enters fullscreen, it needs to tell the tab panel to enter fullscreen to properly appear on top of other panels and controls.

This change also removes the "fullscreen" feature from the quick access controls. There is already a fullscreen icon in the panel toolbar which is available without having to open any menus or special shortcut keys.

Fixes: #2248