Refactor pose handling for OccupancyGrid messages
OccupancyGrid messages with a frame_id different than the currently viewed frame_id in the 3D panel were not rendering with the correct pose. This is because the OG shader was reading pose from `.info.origin` but also using applyPose, which reads the `.pose` field that is being updated per frame. This removes special transform handling for OG messages and unifies on `.pose`.