Remove RandomAccessDataProviderDescriptor concept
Rather than describing a pipeline of data provides abstractly with descriptors, each factory that wants to create a RandomAccessPlayer explicitly creates the DataProviders required for its data source. This allows the factory to create the data provider in a way appropriate for the specific data source rather than requiring all components to adhere to one descriptor specification.

This is a precursor to putting more data providers inside workers.