Add support for recent file
When a file is opened via the open dialog, we record an entry into recents with the FileSystemFileHandle. This entry can later be clicked to re-open the existing file.

On web this feature is restricted to https and may prompt the user to confirm file access before the app is able to read the file again.