Fix url state for data sources using RandomAccessPlayer
A recent change to remove the RandomAccessPlayerDescriptors broke url state handling which relied on a _label member of RandomAccessPlayer.

This is fixed by having data source factories initialize RandomAccessPlayer with appropriate options if providing a url state is important for the specific data source.

Update url state handling to unset data source fields if a data source does not provide state. Data sources which do not provide url state are those which cannot be restored from url state.

Fixes: #2397