Ignore CMD modifier for 3d panel `3` hotkey
The 3d panel sets up a keyboard shortcut on the `3` key. By default the keylister captures events with modifier keys and also prevents default. This makes this keyboard shortcut capture events for `cmd+3` as well as `3`. In chrome, cmd+3 is a shortcut for switching tabs. Since the 3d panel only needs the `3` keylistener, we avoid preventing default when the meta or ctrl key is also pressed.

Fixes: #2371