Reset OpenDialog to start view after open file cancel
When the user clicks "Open file" in the open file dialog and decides not to open a file, we would incorrectly treat this as a no-op to the state of the open dialog.

This change fixes the behavior by setting the view back to start, so future clicks of "Open file" result in launching the file dialog again.

Fixes: #2521