Fix double-click file open on MacOS with no windows
With MacOS it is possible to close all active windows but keep the application open. If this happened, our desktop app would stop handling double-click file open actions since there was no window to handle the action.

This change fixes the behavior by opening a window if there are no windows.

Fixes: #2539