Fix parameter updating
This change fixes two issues. The first is with the Ros1DataSourceFactory. The factory was initializing the Ros1Player using the wrong argument name. The argument name is _hostname_ which matches the form field _id_ configured within the same factory.

The second change fixes JSONInput. JSONInput displays json to the user and allows them to edit the json. To accomplish this, the component has internal state that tracks the latest value of the value prop. It also had logic to update the state if the value was different from the prop. This set state was called in render producing a side effect.

Instead of setting the state in render and tracking the previous prop value with a ref, we track the _editing_ state and set the internal state value only when the user is not editing the component.

Fixes: #2644