Clamp currentTime to start/end for FoxgloveDataPlatformPlayer
When the player is given a start/end range that does not match available coverage, it will clamp the start/end range to a valid coverage range. Prior to this change, this would also unconditionally set the currentTime to the start time even if the currentTime fell within the new valid range. This change clamps the currentTime to the new valid range.

Fixes: #2821