Resubscribe when rosbridge connection is re-established
This change fixes a behavior where the rosbridge connection would re-establish but fail to re-subscribe. This occured due to logic in rosbridge player that would detect that the topics did not change and bail early before subscribing.

This change fixes the logic by adding a `forceUpdate` option to `requestTopics` which forces the topic update to complete even if the topics have not changed. This forceUpdate is used in the "connection" handler when a connection is established.

Fixes: #2866